Changed tests for backend directory structured
